fix(go-rewrite): dedupe parallel-PR helper redeclarations
Wave 2 round-2 PRs (CreateUser, GetUser, ListUsers, AddTenantMember,
ChangeMemberRole, ...) each independently added the same helpers.
When all merged onto main, `go vet ./internal/api/...` failed with
`userToProto redeclared` and `Server.lookupMemberRole already
declared`.

Consolidate into server/go/internal/api/helpers.go (and
helpers_external_test.go for test-only helpers).

Refs #407
